requirements.txt
ultralytics
flask
werkzeug
opencv-python
pillow
pillow-heif
numpy